requirements/base.txt
Django==1.6.1
bpython==0.12
django-braces==1.2.2
django-model-utils==1.4.0
logutils==0.3.3
South==0.8.4
django-social-auth==0.7.28
scrapy==0.20.2
django-celery==3.1.1
beautifulsoup4==4.3.2
selenium==2.39.0
django-stripe-payments==2.0b23